+++ flecsi/execution/test/handle_list.cc
@@ -35,8 +35,8 @@ write_task(data_client_handle_u<mesh_t, ro> mesh,
   auto & context = execution::context_t::instance();
   const auto & map = context.index_map(cells);
   for(auto c : mesh.cells(flecsi::owned)) {
-    fs[0](c) = map.at(c.id());
-    fs[1](c) = 2 * map.at(c.id());
+    fs[0](c) = static_cast<int>(map.at(c.id()));
+    fs[1](c) = static_cast<int>(2 * map.at(c.id()));
   }
 } // task1
